[IR JIT] Playing with JIT, minor updates and minor code cleanup.

+++ src/org/jruby/ir/instructions/specialized/OneFixnumArgNoBlockCallInstr.java
@@ -18,7 +18,7 @@ public class OneFixnumArgNoBlockCallInstr extends CallInstr {
         
         assert getCallArgs().length == 1;
         
-        this.arg1 = ((Fixnum) getCallArgs()[0]).value.longValue();        
+        this.arg1 = ((Fixnum) getCallArgs()[0]).value;
     }
     
     @Override
